+++ Blade Runner/Assets/__Scripts/Core/Core.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Core : MonoBehaviour
+{
+
+}
