+++ Blade Runner/Assets/__Scripts/Core/Core.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 793894b9533d9bf439db86270e0658a4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
