+++ Blade Runner/Assets/__Scripts/Core/CoreComponents.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c3a0ed66e327020408d9da46dee6d368
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
