+++ Blade Runner/Assets/__Scripts/Core/CoreComponents/CoreComponent.cs
@@ -1,0 +1,18 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class CoreComponent : MonoBehaviour
+{
+    protected Core core;
+
+    protected virtual void Awake()
+    {
+        core = transform.parent.GetComponent<Core>();
+
+        if (core == null)
+        {
+            Debug.LogError("There is no Core in the parent");
+        }
+    }
+}
