+++ Blade Runner/Assets/__Scripts/Core/CoreComponents/CoreComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 269e01115012d5f4083a3b7632638648
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
