+++ Blade Runner/Assets/__Scripts/Core/CoreComponents/Movement.cs
@@ -1,0 +1,55 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Movement : CoreComponent
+{
+    private Vector2 workSpace;
+
+    public Vector2 currentVelocity { get; private set; }
+    public Rigidbody2D RB { get; private set; }
+
+    protected override void Awake()
+    {
+        base.Awake();
+
+        RB = GetComponentInParent<Rigidbody2D>();
+    }
+
+    #region Set functions
+    public void SetVelocityX(float velocity)
+    {
+        workSpace.Set(velocity, currentVelocity.y);
+        RB.velocity = workSpace;
+        currentVelocity = workSpace;
+    }
+
+    public void SetVelocityY(float velocity)
+    {
+        workSpace.Set(currentVelocity.x, velocity);
+        RB.velocity = workSpace;
+        currentVelocity = workSpace;
+    }
+
+    public void SetVelocity(float velocity, Vector2 angle, int direction)
+    {
+        angle.Normalize();
+        workSpace.Set(angle.x * velocity * direction, angle.y * velocity);
+        RB.velocity = workSpace;
+        currentVelocity = workSpace;
+    }
+
+    public void SetVelocityZero()
+    {
+        RB.velocity = Vector2.zero;
+        currentVelocity = Vector2.zero;
+    }
+
+    public void SetVelocity(float velocity, Vector2 direction)
+    {
+        workSpace = direction * velocity;
+        RB.velocity = workSpace;
+        currentVelocity = workSpace;
+    }
+    #endregion
+}
