+++ Blade Runner/Assets/__Scripts/Core/CoreComponents/Movement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 07d2e4e34c915f940b90b94e73349feb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
