+++ Blade Runner/Assets/__Scripts/Player/PlayerFiniteStateMachine/Player.cs
@@ -107,43 +107,6 @@ public class Player : MonoBehaviour
     }
     #endregion
 
-    #region Set functions
-    public void SetVelocityX(float velocity)
-    {
-        workSpace.Set(velocity, currentVelocity.y);
-        RB.velocity = workSpace;
-        currentVelocity = workSpace;
-    }
-
-    public void SetVelocityY(float velocity)
-    {
-        workSpace.Set(currentVelocity.x, velocity);
-        RB.velocity = workSpace;
-        currentVelocity = workSpace;
-    }
-
-    public void SetVelocity(float velocity, Vector2 angle, int direction)
-    {
-        angle.Normalize();
-        workSpace.Set(angle.x * velocity * direction, angle.y * velocity);
-        RB.velocity = workSpace;
-        currentVelocity = workSpace;
-    }
-
-    public void SetVelocityZero()
-    {
-        RB.velocity = Vector2.zero;
-        currentVelocity = Vector2.zero;
-    }
-
-    public void SetVelocity(float velocity, Vector2 direction)
-    {
-        workSpace = direction * velocity;
-        RB.velocity = workSpace;
-        currentVelocity = workSpace;
-    }
-    #endregion
-
     #region check functions
     public void CheckIfSchouldFlip(int xinput)
     {
